Add AR Sunglasses virtual try-on example
Extract the generic head-tracking core of the optic blast example
(MediaPipe face landmarker driving the camera, with adaptive smoothing,
velocity prediction, cover-crop FOV matching and nose bridge anchor
correction) into a reusable faceTracking script. OpticBlastTracking now
extends it, keeping only the hand landmarker and ear-touch detection,
and the readiness event is renamed from visor:ready to face:ready.

The new example wears the Khronos glTF sample sunglasses (CC BY 4.0) at
the nose bridge of canonical face space, with two more scripts extracted
from optic blast concepts: headOccluder (a depth-only ellipsoid that
hides the temple arms behind the head) and lensTint (converts the glTF
transmission lens materials to alpha blending, since refraction cannot
sample the DOM video behind the transparent canvas, and retints the
lenses from a row of swatches on the page).
